- Closes https://github.com/ppy/osu-framework/issues/5084
- Closes https://github.com/ppy/osu-framework/issues/5499
- Closes https://github.com/ppy/osu-framework/issues/2073
- Closes https://github.com/ppy/osu/issues/8580
- Supersedes / closes https://github.com/ppy/osu-framework/pull/5507

You might ask why I'm bothering to try this now. Well, this came up when I wanted to use text flow for BSS purposes (and forced me to hack around it), and came up again in https://github.com/ppy/osu/pull/31970, and I'm sick of it.

The actual fix is taken verbatim from https://github.com/ppy/osu-framework/pull/5507, it's just restructured using the idea of a single nested flow taken from https://github.com/ppy/osu-framework/pull/5507#discussion_r1023426911.

Game side changes available for preview at https://github.com/ppy/osu/compare/master...bdach:osu:text-flow-game-side-adjustments?expand=1

## Benchmarks

### Before

|  Method |     Mean |     Error |    StdDev |   Median |    Gen0 |    Gen1 | Allocated |
|-------- |---------:|----------:|----------:|---------:|--------:|--------:|----------:|
| SetText | 1.157 ms | 0.0565 ms | 0.1666 ms | 1.094 ms | 99.6094 | 85.9375 |    1.2 MB |

### After

|  Method |     Mean |     Error |    StdDev |    Gen0 |    Gen1 | Allocated |
|-------- |---------:|----------:|----------:|--------:|--------:|----------:|
| SetText | 1.169 ms | 0.0536 ms | 0.1580 ms | 89.8438 | 74.2188 |   1.09 MB |

## Breaking changes

### `TextFlowContainer` no longer inherits from `FillFlowContainer` and instead is a `CompositeDrawable`

All methods and properties that consumers should be using (and additionally some that they arguably shouldn't) remain accessible.